+++ include/render/GpuDiffusePathLoopLaunch.h
@@ -8,7 +8,7 @@
 #include <vector>
 
 namespace render {
-  inline constexpr std::uint32_t gpuDiffusePathLoopLaunchLayoutVersion = 14u;
+  inline constexpr std::uint32_t gpuDiffusePathLoopLaunchLayoutVersion = 15u;
 
   struct alignas(16) GpuDiffusePathLoopLaunchParameters {
     std::uint32_t layoutVersion{gpuDiffusePathLoopLaunchLayoutVersion};
@@ -67,9 +67,12 @@ namespace render {
     std::uint32_t displayResolveTonemap{
       static_cast<std::uint32_t>(GpuDisplayResolveTonemap::Linear)};
     std::uint32_t captureMetrics{1};
-    std::uint32_t reserved3{0};
+    std::uint32_t primaryPathMotionMode{gpuPrimaryPathMotionModeOriginDelta};
     std::array<float, 4> primaryPathOrigin{};
     std::array<float, 4> primaryPathMotionOriginDelta{};
+    std::array<float, 4> primaryPathMotionTarget{};
+    std::array<float, 4> primaryPathMotionTargetDelta{};
+    std::array<float, 4> primaryPathMotionParameters{};
     std::array<float, 4> primaryPathTopLeft{};
     std::array<float, 4> primaryPathRight{};
     std::array<float, 4> primaryPathDown{};